% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln.R
\name{Binomial-logit-normal}
\alias{Binomial-logit-normal}
\alias{bln}
\alias{dbln}
\alias{pbln}
\alias{qbln}
\alias{rbln}
\title{The Binomial-Logit-Normal Distribution}
\usage{
dbln(x, size, mean = 0, sd = 1)

pbln(q, size, mean = 0, sd = 1)

qbln(...)

rbln(n, size, mean = 0, sd = 1)
}
\arguments{
\item{x, q}{A vector of quantiles}

\item{size}{Number of trials}

\item{mean}{A vector of means}

\item{sd}{A vector of standard deviations}

\item{n}{Number of observations, if \code{length(x = n) > 1},
the length is taken to be the number required}

\item{p}{A vector of probabilities}
}
\value{
\code{dbln} gives the density

\code{pbln} gives the distribution function

\code{qbln} gives the quantile function

\code{rbln} generates random deviates
}
\description{
The Binomial-Logit-Normal Distribution
}
