% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_pmn.R
\name{bln_b_pmn}
\alias{bln_b_pmn}
\title{Calculate and evaluate the index for the microbial biological activity in the Netherlands}
\usage{
bln_b_pmn(B_LU_BRP, B_SOILTYPE_AGR, A_N_PMN)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_N_PMN}{(numeric) The potentially mineralizable N pool (mg N / kg soil)}
}
\value{
the normalized potentially mineralizable Nitrogen pool (mg N / kg), a numeric value, converted to an OSI score.
}
\description{
This function assesses the microbial biological activity (of microbes and fungi) via the Potentially Mineralizable N pool, also called PMN (or SoilLife by Eurofins in the past).
}
\examples{
bln_b_pmn(B_LU_BRP = 256, B_SOILTYPE_AGR = 'dekzand', A_N_PMN = 125)
bln_b_pmn(c(256,1027), c('dekzand','rivierklei'), c(125,45))

}
