% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_gw_waterbalance.R
\name{bln_bbwp_bw}
\alias{bln_bbwp_bw}
\title{Function to calculate and evaluate the water buffering capacity of soils in view of water retention in soil}
\usage{
bln_bbwp_bw(
  ID,
  B_LU_BRP,
  B_HELP_WENR,
  B_GWL_CLASS,
  B_AREA_DROUGHT,
  A_CLAY_MI,
  A_SAND_MI,
  A_SILT_MI,
  A_SOM_LOI,
  penalty = TRUE
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code}

\item{B_HELP_WENR}{(character) The soil type abbreviation, derived from 1:50.000 soil map}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_AREA_DROUGHT}{(boolean) is the field located in an area with high risks for water deficiencies (options: TRUE or FALSE)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{penalty}{(boolean) the option to apply a penalty for high risk BBWP field indicators. Default is TRUE.}
}
\description{
This function gives the WB score of the BBWP framework
}
\examples{
bln_bbwp_bw(
ID = 15,
B_LU_BRP = c(233,259,2014,308),
B_HELP_WENR = c('gMn53C', 'Mn25C','U0102nr108','nSn13A'),
B_GWL_CLASS = rep('GtVI',4),
B_AREA_DROUGHT = rep(TRUE,4),
A_CLAY_MI = rep(20,4),
A_SAND_MI = rep(15,4),
A_SILT_MI = rep(10,4),
A_SOM_LOI = c(2,3,5,8),
penalty = TRUE
)

}
