% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_gw_bbwp_nitrate.R
\name{bln_bbwp_ngw}
\alias{bln_bbwp_ngw}
\title{Function to calculate and evaluate the N buffering capacity of soils in view of water purification for groundwater quality}
\usage{
bln_bbwp_ngw(
  ID,
  B_LU_BRP,
  B_SOILTYPE_AGR,
  B_SC_WENR,
  B_AER_CBS,
  B_GWP,
  B_GWL_CLASS,
  A_SOM_LOI,
  A_N_RT,
  B_N_RT = NA_real_,
  B_N_RT_SD = NA_real_,
  penalty = TRUE
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_SC_WENR}{(integer) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006). Options include: 1,2,3,4,5,10,11,401,901 and 902.}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{B_GWP}{(boolean) is the field located in a groundwater protected area (options: TRUE or FALSE)}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{B_N_RT}{(numeric) The mean organic nitrogen content of the soil in the LSW region (mg N / kg). Optional.}

\item{B_N_RT_SD}{(numeric) The variance in organic nitrogen content of the soil in the LSW region (standard deviation) (mg N / kg). Optional.}

\item{penalty}{(boolean) the option to apply a penalty for high risk BBWP field indicators. Default is TRUE.}
}
\description{
This function gives the NGW score of the BBWP framework
}
