% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_sw_bbwp_phosphate.R
\name{bln_bbwp_psw}
\alias{bln_bbwp_psw}
\title{Function to calculate and evaluate the P buffering capacity of soils in view of water purification for surface quality}
\usage{
bln_bbwp_psw(
  ID,
  B_LU_BRP,
  B_SC_WENR,
  B_AER_CBS,
  B_GWL_CLASS,
  B_SLOPE_DEGREE,
  A_P_CC,
  A_P_SG,
  A_AL_OX,
  A_FE_OX,
  D_RO_R,
  D_SA_W,
  B_RO_R = NA_real_,
  B_RO_R_SD = NA_real_,
  B_P_CC = NA_real_,
  B_P_CC_SD = NA_real_,
  B_P_SG = NA_real_,
  B_P_SG_SD = NA_real_,
  B_AL_OX = NA_real_,
  B_AL_OX_SD = NA_real_,
  B_FE_OX = NA_real_,
  B_FE_OX_SD = NA_real_,
  B_CT_PSW,
  B_CT_PSW_MAX = 0.5,
  penalty = TRUE
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code}

\item{B_SC_WENR}{(integer) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006). Options include: 1,2,3,4,5,10,11,401,901 and 902.}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}

\item{A_P_CC}{(numeric) The plant available P content, measured via 0.01M CaCl2 extraction (mg / kg)}

\item{A_P_SG}{(numeric) The P-saturation index (\%)}

\item{A_AL_OX}{(numeric) The iron content of soil (mmol / kg)}

\item{A_FE_OX}{(numeric) The aluminium content of soil (mmol / kg)}

\item{D_RO_R}{(numeric) The risk that surface water runs off the parcel}

\item{D_SA_W}{(numeric) The wet perimeter index of the field, fraction that field is surrounded by water}

\item{B_RO_R}{(numeric) The mean D_RO_R of the soil in the LSW region. Optional.}

\item{B_RO_R_SD}{(numeric) The variance in D_RO_R in the LSW region (standard deviation). Optional.}

\item{B_P_CC}{(numeric) The mean P-CaCl2 content of the soil in the LSW region (mg P / kg). Optional.}

\item{B_P_CC_SD}{(numeric) The variance in P-CaCl2 content  of the soil in the LSW region (standard deviation) (mg P / kg). Optional.}

\item{B_P_SG}{(numeric) The mean P saturation degree of the soil in the LSW region (mg N / kg). Optional.}

\item{B_P_SG_SD}{(numeric) The variance in P saturation degree of the soil in the LSW region (standard deviation) (\%). Optional.}

\item{B_AL_OX}{(numeric) The mean Al-oxide content of the soil in the LSW region (\%). Optional.}

\item{B_AL_OX_SD}{(numeric) The variance in Al-oxide content of the soil in the LSW region (standard deviation) (mmol+ / kg). Optional.}

\item{B_FE_OX}{(numeric) The mean Fe-oxide content of the soil in the LSW region (mmol+ / kg). Optional.}

\item{B_FE_OX_SD}{(numeric) The variance in Fe-oxide content of the soil in the LSW region (standard deviation) (mmol+ / kg). Optional.}

\item{B_CT_PSW}{(numeric) the critical target for required reduction in P loss from agriculture (kg P / ha) to reach targets of KRW}

\item{B_CT_PSW_MAX}{(numeric) the max critical target for P reduction loss (kg P / ha)}

\item{penalty}{(boolean) the option to apply a penalty for high risk BBWP field indicators. Default is TRUE.}
}
\description{
This function gives the PSW score of the BBWP framework
}
