% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_cec.R
\name{bln_c_cec}
\alias{bln_c_cec}
\title{Calculate and evaluate the soil fertility based on the CEC}
\usage{
bln_c_cec(A_CEC_CO)
}
\arguments{
\item{A_CEC_CO}{(numeric) The cation exchange capacity (mmol+ / kg)}
}
\value{
The capacity of the soil to buffer cations. A numeric value.
}
\description{
This function calculates the capacity of the soil to buffer cations
}
