% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_nitrogen.R
\name{bln_c_nitrogen}
\alias{bln_c_nitrogen}
\title{Calculate the soil nitrogen supplying capacity in the Netherlands}
\usage{
bln_c_nitrogen(B_LU_BRP, B_SOILTYPE_AGR, A_SOM_LOI, A_N_RT)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}
}
\value{
The capacity of the soil to supply nitrogen (kg N / ha / yr). A numeric value, converted to a OSI score.
}
\description{
This function calculates the NLV (nitrogen producing capacity) for the soil
}
\examples{
bln_c_nitrogen(B_LU_BRP = 256, B_SOILTYPE_AGR = 'dekzand',A_SOM_LOI = 4.5,A_N_RT = 2500)
bln_c_nitrogen(1019,'dekzand',5.5,2315)

}
