% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_ph.R
\name{bln_c_ph}
\alias{bln_c_ph}
\title{Calculate the distance for target for soil pH in view of the BLN production function}
\usage{
bln_c_ph(ID, B_LU_BRP, B_SOILTYPE_AGR, A_SOM_LOI, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The organic matter content of soil in percentage}

\item{A_CLAY_MI}{(numeric) The percentage A_CLAY_MI present in the soil}

\item{A_PH_CC}{(numeric) The pH-CaCl2 of the soil}
}
\value{
The bln indicator for the soil pH
}
\description{
This functions evaluates the difference between the measured pH and the optimal pH according to the Bemestingsadvies
}
\references{
Handboek Bodem en Bemesting tabel 5.1, 5.2 en 5.3
}
