% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_sulfur.R
\name{bln_c_sulfur}
\alias{bln_c_sulfur}
\title{Calculate the capacity of soils to supply sulfur for the BLN production function}
\usage{
bln_c_sulfur(B_LU_BRP, B_SOILTYPE_AGR, B_AER_CBS, A_SOM_LOI, A_S_RT)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (in percent)}

\item{A_S_RT}{(numeric) The total Sulpher content of the soil (in mg S per kg)}
}
\value{
The capacity of the soil to supply Sulfur (kg S / ha / yr). A numeric value.
}
\description{
This function calculates a S-balance given the SLV (sulfur supplying capacity) of a soil
}
