% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_calc_psp.R
\name{bln_calc_psp}
\alias{bln_calc_psp}
\title{Calculate the precipitation surplus}
\usage{
bln_calc_psp(ID, B_LU_BRP, M_GREEN)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional, options: TRUE, FALSE)}
}
\value{
The estimated precipitation surplus (in mm / ha) depending on averaged precipitation and evaporation. A numeric value.
}
\description{
This function calculates the precipitation surplus (in mm / ha) given the crop rotation plan. Is identical to OBIC function calc_psp but much faster to facilitate national and regional calculations
}
\examples{
bln_calc_psp(ID = 1,B_LU_BRP = 265, M_GREEN = TRUE)
bln_calc_psp(ID = c(1,2,1,2),B_LU_BRP = c(265,1019,265,1019), M_GREEN = rep(TRUE,4))

}
