% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_cbalance.R
\name{bln_clim_cbalance}
\alias{bln_clim_cbalance}
\title{Function to calculate and evaluate a default SOM balance}
\usage{
bln_clim_cbalance(
  ID,
  B_LU_BRP,
  A_SOM_LOI,
  A_P_AL,
  A_P_WA,
  M_COMPOST = 0,
  M_GREEN = FALSE
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{A_P_AL}{(numeric) The P-AL content of the soil}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)}

\item{M_COMPOST}{(numeric) The frequency that compost is applied (optional, every x years)}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional)}
}
\description{
Function to calculate and evaluate a default SOM balance
}
