% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_rothc.R
\name{bln_clim_rothc}
\alias{bln_clim_rothc}
\title{Function to evaluate the carbon saturation via RothC simulation}
\usage{
bln_clim_rothc(
  ID,
  B_LU_BRP,
  B_GWL_GLG,
  A_SOM_LOI,
  A_CLAY_MI,
  quiet = FALSE,
  mc = TRUE
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{B_GWL_GLG}{(numeric) The lowest groundwater level averaged over the most dry periods in 8 years in cm below ground level}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{A_CLAY_MI}{(numeric) value for the clay content of the soil}

\item{quiet}{(boolean) showing progress bar for calculation RothC C-saturation for each field}

\item{mc}{(boolean) option to run rothc in parallel on multicores}
}
\value{
An indicator value for carbon saturation ranging from 0 to 1 where 1
is optimal.
}
\description{
Function to evaluate the carbon saturation via RothC simulation
}
\examples{
bln_clim_rothc(
ID = 15,
A_SOM_LOI = 7.92,
B_GWL_GLG = 75,
A_CLAY_MI = 3.5,
B_LU_BRP = 3732,
quiet = TRUE,
mc = FALSE
)

}
