% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_somers.R
\name{bln_clim_somers}
\alias{bln_clim_somers}
\title{Function to evaluate the carbon saturation via SOMERS simulation on peat soils}
\usage{
bln_clim_somers(
  ID,
  B_SOILTYPE_AGR,
  A_SOM_LOI,
  B_SOMERS_BC,
  B_DRAIN_SP,
  B_DRAIN_WP,
  B_DRAIN_SP_CHANGE = 0.2
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{B_SOMERS_BC}{(integer) The base combination of SOMERS peat soil classification (varies between 1 and 290)}

\item{B_DRAIN_SP}{(numeric) the drooglegging of a field in summer (difference field height and ditch level, in meters)}

\item{B_DRAIN_WP}{(numeric) the drooglegging of a field in winter (difference field height and ditch level, in meters)}

\item{B_DRAIN_SP_CHANGE}{(numeric) the decrease in drooglegging of a field in summer (in meters). Allowed decrease varies from 0 to 0.5m.}
}
\value{
Returns an indicator score (between 0 and 1) for carbon saturation of
 peat and peaty (A_SOM_LOI > 20%) soils where 1 means that the soil is optimally saturated.
}
\description{
Function to evaluate the carbon saturation via SOMERS simulation on peat soils
}
\examples{
bln_clim_somers(
ID = 1,
B_SOILTYPE_AGR = c('veen', 'dalgrond'),
A_SOM_LOI = c(25,21),
B_SOMERS_BC = c(25,25),
B_DRAIN_SP = c(0.63, 0.63),
B_DRAIN_WP = c(0.49, 0.49),
B_DRAIN_SP_CHANGE = c(0.2, 0.2)
)

}
