% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_evaluate.R
\name{bln_evaluate_logistic_gaus_down}
\alias{bln_evaluate_logistic_gaus_down}
\title{Evaluate logistically initially and gausian eventually}
\usage{
bln_evaluate_logistic_gaus_down(x, b, x0, v, optimum, optimum_ofset = 1.5)
}
\arguments{
\item{x}{(numeric) The values of a calc function to be converted to an evaluation}

\item{b}{(numeric) The growth rate}

\item{x0}{(numeric) The offset of the x-axis}

\item{v}{(numeric) Affects the growth rate near the maximum}

\item{optimum}{(numeric) Point where x has approximaltly reached the optimal value}

\item{optimum_ofset}{(numeric) Multiplication factor to determine at what point gaussian evaluation should be applied}
}
\value{
A transformed variable after applying a logistic evaluation function. A numeric value.
}
\description{
This function evaluates the calculated values from an indicator using a general logistic function
}
\examples{
bln_evaluate_logistic_gaus_down(x = 5, b = 2, x0 = 3, v = 2.6, optimum = 1)
bln_evaluate_logistic_gaus_down(x = c(0.1,0.5,1.5,3.5), b = 2, x0 = 3, v = 2.6, optimum = 0.5)

}
\references{
\url{https://en.wikipedia.org/wiki/Generalised_logistic_function}
\url{https://en.wikipedia.org/wiki/Gaussian_function}
}
