% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_farm_hf}
\alias{bln_farm_hf}
\title{An example dataset with all BLN input prepared for de Marke farm}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 330 rows and 78 columns.
}
\usage{
bln_farm_hf
}
\description{
This table containsall input data for an example farm to run the BLN
}
\details{
\describe{
  \item{c1}{}
  \item{c2}{}
  \item{c3}{}
  \item{c4}{}
  \item{c5}{}
  \item{c6}{}
}
}
\keyword{datasets}
