% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_field.R
\name{bln_field_dt}
\alias{bln_field_dt}
\title{Calculate the BLN score for one field using a data.table as input}
\usage{
bln_field_dt(dt, LSW = NULL, output = "all", runrothc = FALSE, mc = FALSE)
}
\arguments{
\item{dt}{(data.table) A data.table with all input required to calculate BLN on field level}

\item{LSW}{(data.table) The averaged soil properties (mean and sd) per Local Surface Water. Can be derived from bbwp_lsw_properties.}

\item{output}{(character) An optional argument to select output: scores, indicators, or all. (default = all)}

\item{runrothc}{(boolean) An argument to switch off rothc calculations due to running time (default: FALSE)}

\item{mc}{(boolean) option to run rothc in parallel on multicores}
}
\description{
This functions wraps the functions of the BLN2 into one main function to calculate the soil quality score for a single field. Whereas the function `bln_field` requires all variables as separate inputs, the function `bln_field_dt` allows one to send in a data.table
}
