% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_helpers.R
\name{bln_format_aer}
\alias{bln_format_aer}
\title{Convert possible B_AER_CBS values to standardized values}
\usage{
bln_format_aer(B_AER_CBS, type = "name")
}
\arguments{
\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{type}{(character) the desired output: the AER_CBS code ("code") or the AER_CBS name ("name")}
}
\value{
A standardized B_AER_CBS value as required for the OBIC functions. A character string.
}
\description{
This function formats information of Agricultural Economic Region so it can be understood by other OBIC functions
}
\examples{
bln_format_aer(c("LG13","LG12"))
bln_format_aer(c("LG13","LG12",'Rivierengebied'))

}
