% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_format_gtclass.R
\name{bln_format_gtclass}
\alias{bln_format_gtclass}
\title{Determine groundwater table class based on highest and lowest mean groundwater levels}
\usage{
bln_format_gtclass(B_GWL_GHG, B_GWL_GLG)
}
\arguments{
\item{B_GWL_GHG}{(numeric) The average highest groundwater level in cm below field level (Gemiddeld Hoogste Grondwaterstand)}

\item{B_GWL_GLG}{(numeric) The average lowest groundwater level in cm below field level (Gemiddeld Laagste Grondwaterstand)}
}
\value{
A standardized B_GWL_CLASS value as required for BLN functions. A character string.
}
\description{
Determine groundwater table class based on highest and lowest mean groundwater levels
}
\examples{
bln_format_gtclass(B_GWL_GHG = 35, B_GWL_GLG = 55)
bln_format_gtclass(B_GWL_GHG = 70, B_GWL_GLG = 155)

}
