% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_input_description}
\alias{bln_input_description}
\title{A dataset with all input variables required by the BLN package}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 5 columns.
}
\usage{
bln_input_description
}
\description{
This table contains all input parameters needed to run the BLN package
}
\details{
\describe{
  \item{code}{}
  \item{parameter}{}
  \item{data_type}{}
  \item{value_min}{}
  \item{value_max}{}
}
}
\keyword{datasets}
