% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_lsw_farm_hf}
\alias{bln_lsw_farm_hf}
\title{An example dataset with all LSW properties related to the fields of the Marke farm}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 27 columns.
}
\usage{
bln_lsw_farm_hf
}
\description{
This table contains the LSW properties for the fields belonging to the Marke farm
}
\details{
\describe{
  \item{c1}{}
  \item{c2}{}
  \item{c3}{}
  \item{c4}{}
  \item{c5}{}
  \item{c6}{}
}
}
\keyword{datasets}
