% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_nut_nitrogen.R
\name{bln_nut_nitrogen}
\alias{bln_nut_nitrogen}
\title{Function to calculate and evaluate the nitrogen use efficiency in view of the soils' function to improve nutrient recycling}
\usage{
bln_nut_nitrogen(ID, B_LU_BRP, B_SOILTYPE_AGR, A_SOM_LOI, A_N_RT)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code}

\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}
}
\description{
Function to calculate and evaluate the nitrogen use efficiency in view of the soils' function to improve nutrient recycling
}
