% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_nut_nue_bbwp.R
\name{bln_nut_nue}
\alias{bln_nut_nue}
\title{Function to calculate and evaluate the NUE}
\usage{
bln_nut_nue(
  B_LU_BRP,
  B_HELP_WENR,
  B_GWL_CLASS,
  A_P_AL,
  A_P_CC,
  A_P_WA,
  A_N_RT,
  B_N_RT = NA_real_,
  B_N_RT_SD = NA_real_,
  penalty = TRUE
)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code}

\item{B_HELP_WENR}{(character) The soil type abbreviation, derived from 1:50.000 soil map}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_P_AL}{(numeric) The P-AL content of the soil}

\item{A_P_CC}{(numeric) The plant available P content, measured via 0.01M CaCl2 extraction (mg / kg)}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{B_N_RT}{(numeric) The mean organic nitrogen content of the soil in the LSW region (mg N / kg). Optional.}

\item{B_N_RT_SD}{(numeric) The variance in organic nitrogen content of the soil in the LSW region (standard deviation) (mg N / kg). Optional.}

\item{penalty}{(boolean) the option to apply a penalty for high risk BBWP field indicators. Default is TRUE.}
}
\description{
This is the NUE calculation as being used in the BBWP framework
}
