% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_nut_phosphorus.R
\name{bln_nut_phosphorus}
\alias{bln_nut_phosphorus}
\title{Function to calculate and evaluate the phosphorus use efficiency in view of the soils' function to improve nutrient recycling}
\usage{
bln_nut_phosphorus(
  B_LU_BRP,
  A_P_AL = NA_real_,
  A_P_CC = NA_real_,
  A_P_WA = NA_real_
)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with ammonium lactate}

\item{A_P_CC}{(numeric) The P-content of the soil extracted with CaCl2}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water}
}
\description{
Function to calculate and evaluate the phosphorus use efficiency in view of the soils' function to improve nutrient recycling
}
\examples{
bln_nut_phosphorus(B_LU_BRP = 265, A_P_AL = 45, A_P_CC = 2.5)
bln_nut_phosphorus(B_LU_BRP = c(265,1019),A_P_AL = c(35,54),A_P_CC = c(2.5,4.5), A_P_WA = c(35,65))

}
