% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_output_description}
\alias{bln_output_description}
\title{A dataset with all output variables generated by the BLN package}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 49 rows and 2 columns.
}
\usage{
bln_output_description
}
\description{
This table contains all soil indicators and soil quality scores for the BLN package
}
\details{
\describe{
  \item{code}{}
  \item{parameter}{}
}
}
\keyword{datasets}
