% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_crumbleability.R
\name{bln_p_crumbleability}
\alias{bln_p_crumbleability}
\title{Calculate and evaluate the crumbleability in view of the soils production function}
\usage{
bln_p_crumbleability(B_LU_BRP, A_SOM_LOI, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{A_SOM_LOI}{(numeric) The organic matter content of soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_CC}{(numeric) The pH of the soil, measured in 0.01M CaCl2}
}
\value{
The crumbleability index of a soil, a measure for a physical soil property. A numeric value.
}
\description{
This function calculates the crumbleability.
}
