% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_density.R
\name{bln_p_density}
\alias{bln_p_density}
\title{Calculate the bulk density}
\usage{
bln_p_density(A_SOM_LOI, A_CLAY_MI, A_DENSITY_SA = NA_real_)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_DENSITY_SA}{(numeric) The bulk density of the soil (kg/m3)}
}
\value{
The bulk density of an arable soil (kg / m3) evaluated given the maximum density that limit the root penetration.
}
\description{
This pedotransfer function calculates the bulk density of the soil based on texture and organic matter
}
\examples{
bln_p_density(A_SOM_LOI = 6.5, A_CLAY_MI = 28)
bln_p_density(A_SOM_LOI = 3.5, A_CLAY_MI = 2)
bln_p_density(A_SOM_LOI = c(3.5,8.5),A_CLAY_MI = c(2,28))

}
