% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_sealing.R
\name{bln_p_sealing}
\alias{bln_p_sealing}
\title{Calculate and evaluate the soil sealing risk}
\usage{
bln_p_sealing(B_LU_BRP, A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code}

\item{A_SOM_LOI}{(numeric) The organic matter content of soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}
}
\value{
The risk of soil sealing as affected by the soil organic matter and clay content. A numeric value.
}
\description{
This function calculates the risks of soil sealing.
}
