% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_prod_winderosion.R
\name{bln_p_winderosion}
\alias{bln_p_winderosion}
\title{Calculate and evaluate the indicator for wind erodibility in view of food production}
\usage{
bln_p_winderosion(B_LU_BRP, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}
}
\value{
The vulnerability of the soil for wind erosion. A numeric value.
}
\description{
This function calculates and evaluates the risk for wind erodibility of soils, derived from Van Kerckhoven et al. (2009) and Ros & Bussink (2013)
}
