% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_parms}
\alias{bln_parms}
\title{Table with BLN parameters being used in the package}
\format{
A data.table with x rows and x columns:
\describe{
  \item{bln_parm_id}{the parameter id}
  \item{bln_parm_name}{the name of the parameter}
  \item{bln_parm_type}{the type of the parameter. Options: measurement, field property}
  \item{bln_parm_description}{a short description of the parameters}
  \item{bln_parm_unit}{the unit of the parameter}
  \item{bln_parm_min}{the maximum allowed value for the parameter}
  \item{bln_parm_max}{the minimum allowed value for the parameter}
  \item{bln_parm_data_type}{the data type of the parameter: numeric, character or boolean}
  \item{bln_parm_enum}{does the parameter have predefined options}
  \item{bln_parm_options}{allowed options for the parameteer}
}
}
\usage{
bln_parms
}
\description{
This table contains all parameters being used in the BLN package to calculate the soil quality index.
}
\keyword{datasets}
