% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_rothc_sim.R
\name{bln_rothc}
\alias{bln_rothc}
\title{RothC calculation motor}
\usage{
bln_rothc(time, y, parms)
}
\arguments{
\item{time}{time to print output}

\item{y}{state variables to run RothC}

\item{parms}{parameters to run RothC, includes rate modifying functions and constants R1, abc, and d}
}
\description{
Set of differential equations to calculate the evolutotion of Soil Organic Carbon using the RothC model.
}
