% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_rothc_events.R
\name{bln_rothc_event_amendment}
\alias{bln_rothc_event_amendment}
\title{Calculate the monthly timing of carbon inputs for different fertilizer strategies for the Netherlands}
\usage{
bln_rothc_event_amendment(crops, amendment = NULL)
}
\arguments{
\item{crops}{(data.table) Table with crop rotation, year and potential Carbon inputs.}

\item{amendment}{(data.table) A table with the following column names: year, month, cin_tot, cin_hum, cin_dpm, cin_rpm and the fraction eoc over p (fr_eoc_p). Month is optional.}
}
\description{
This function calculates the timing of carbon inputs (kg C per ha) based on type of organic matter amendments and land use.
}
\details{
This function increases temporal detail for time series of C inputs of organic amendments.
The inputs for organic amendments are organised in the data.table amendment, where the carbon inputs has the unit kg C / ha.

The output is an EVENT object.
}
