% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_rothc_events.R
\name{bln_rothc_event_crop}
\alias{bln_rothc_event_crop}
\title{Calculate the crop rotation related C inputs of a field on monthly basis for the Netherlands}
\usage{
bln_rothc_event_crop(crops, A_CLAY_MI)
}
\arguments{
\item{crops}{(data.table) Table with crop rotation, crop management measures, year and potential Carbon inputs.}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\value{
a data table with columns time, var, value, and method
}
\description{
This function determines how much Carbon enters the soil throughout the year given the crop rotation plan.
}
\details{
The crops data.table is typically the output of \link{bln_rothc_input_crop}
}
\examples{
 crops <- data.table::data.table(
   year = c(1, 2, 3, 4),
   B_LU = c("nl_2015", "nl_233", "nl_256", "nl_233"),
   cin_crop_dpm = c(1200, 1520, 468, 1517),
   cin_corp_rpm = c(782, 1127, 312, 1127),
   cin_res_dpm = rep(0, 4),
   cin_res_repm = rep(0, 4),
   M_GREEN_TIMING = c("october", "never", "never", "never"),
   M_IRRIGATION = rep(FALSE, 4),
   M_CROPRESIDUE = rep(FALSE, 4),
   M_RENEWAL = rep(FALSE, 4),
   cf_yield = rep(1, 4))

 bln_rothc_event_crop(crops = crops, A_CLAY_MI = 5)

}
