% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_rothc_inputs.R
\name{bln_rothc_input_amendment}
\alias{bln_rothc_input_amendment}
\title{Do check and expand amendment input table for RothC for the Netherlands}
\usage{
bln_rothc_input_amendment(dt = NULL, B_LU_BRP = NULL)
}
\arguments{
\item{dt}{(data.table) Table with amendments and amendment properties for Carbon input.}

\item{B_LU_BRP}{(numeric) The crop code}
}
\value{
A data table with columns p_name, year, cin_tot, cin_hum, cin_dpm, cin_rpm, and fr_eoc_p
}
\description{
Helper function to check the content and format of the amendment input table.
}
\details{
The amendments table used as input for carbon modelling requires at minimum data on effective organic matter inputs and related year.
This helper function assists the checking and controlling of amendments properties involved.

To run this function, the dt requires as input:"P_NAME", "year","month","P_OM","P_HC","p_p2o5", and "P_DOSE"
if dt is NULL, then the amendment input will be prepared using function \link{rothc_scenario} using scenario 'BAU'
}
\examples{
# example without dt
bln_rothc_input_amendment(B_LU_BRP = 256)

# example with dt
dt.in <- data.table::data.table(
            P_NAME = "cattle_slurry",
            year = 1,
            month = 3,
            P_OM = 800,
            P_HC = 0.25,
            p_p2o5 = 20,
            P_DOSE = 15)
bln_rothc_input_amendment(dt = dt.in, B_LU_BRP = 256)

}
