% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_rothc_inputs.R
\name{bln_rothc_input_rmf}
\alias{bln_rothc_input_rmf}
\title{Estimate the rate modifying factors for the RothC modelling}
\usage{
bln_rothc_input_rmf(dt = NULL, B_LU_BRP = NULL, A_CLAY_MI, simyears)
}
\arguments{
\item{dt}{(data.table) Table with crop rotation and related crop properties for Carbon input.}

\item{B_LU_BRP}{(numeric) The crop code}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{simyears}{(numeric) Amount of years for which the simulation should run, default: 50 years}
}
\description{
This function prepares the rate modifying factors for RothC given crop cover
}
\details{
To run this function, the dt requires as input: B_LU (a crop id), B_LU_NAME (a crop name, optional), B_LU_EOM (the effective organic matter content, kg/ha), B_LU_EOM_RESIDUE (the effective organic matter content for crop residues, kg/ha), and the B_LU_HC (the humification coeffient,-).
if dt is NULL, then the crop input will be prepared using function \link{rothc_scenario} using scenario 'BAU'
}
