% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_scen_croprotation}
\alias{bln_scen_croprotation}
\title{A dataset with potential crop rotation scenarios per agricultural economic region}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 807 rows and 5 columns.
}
\usage{
bln_scen_croprotation
}
\description{
This table contains a series of potential crop rotations
}
\details{
\describe{
  \item{b_aer_cbs}{}
  \item{soiltype}{}
  \item{b_lu_brp}{}
  \item{scen}{}
  \item{year}{}
}
}
\keyword{datasets}
