% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_soiltype}
\alias{bln_soiltype}
\title{Table with country specific soil types being used to evaluate soil quality}
\format{
A data.table with 9 rows and 4 columns:
\describe{
  \item{bln_soil_id}{the soil category id}
  \item{bln_country}{the name of the country where the soil category is used}
  \item{bln_soil_cat1}{soil category 1, might be in the language of the country}
  \item{bln_soil_cat2}{soil category 2, in english}
}
}
\usage{
bln_soiltype
}
\description{
This table contains categories of soil types being used to evaluate soil quality. Categories vary per country.
}
\keyword{datasets}
