% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_somers}
\alias{bln_somers}
\title{A dataset with meta-model results from SOMERS per unique base combination peat soil}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9570 rows and 6 columns.
}
\usage{
bln_somers
}
\description{
This table contains predicted CO2 emissions for various scenarios for predefined peat soil types
}
\details{
\describe{
  \item{c1}{}
  \item{c2}{}
  \item{c3}{}
  \item{c4}{}
  \item{c5}{}
  \item{c6}{}
}
}
\keyword{datasets}
