% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_thresholds}
\alias{bln_thresholds}
\title{Table with country specific threshold values to evaluate soilquality'}
\format{
A data.table with x rows and 12 columns:
\describe{
  \item{bln_trh_id}{the threshold category id}
  \item{bln_country}{the country name where the threshold coefficients apply}
  \item{bln_esd}{the soil ecosystem service for which the threshold applies}
  \item{bln_category}{the soil quality category: chemical, biological, physical, other}
  \item{bln_indicator}{the bln indicator for which the threshold applies}
  \item{bln_indicator_name}{the bln indicator name}
  \item{bln_threshold_cropcat}{the bln crop category for which specific thresholds apply. if empy, then the threshold is generally applicable}
  \item{bln_threshold_soilcat}{the bln soil category for which specific thresholds apply. if empy, then the threshold is generally applicable}
  \item{bln_scoringtype}{the soil evaluation function to score the soil function. Options: parabolic, logistic, linear}
  \item{bln_st_c1}{the first coefficient of the scoring function selected}
  \item{bln_st_c2}{the second coefficient of the scoring function selected}
  \item{bln_st_c3}{the third coefficient of the scoring function selected}
}
}
\usage{
bln_thresholds
}
\description{
This table contains the evaluation coefficients to convert a soil function value into a soil quality score ranging from 0 to 1.
The scoring function migh vary per country, indicator and subcatogries defined by variable categories of soil and crop types.
}
\keyword{datasets}
