% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_variable_grouping}
\alias{bln_variable_grouping}
\title{Grouping of indicators and BLN functions used for aggregation}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 36 rows and 4 columns.
}
\usage{
bln_variable_grouping
}
\description{
\describe{
  \item{bln_function}{Name of a function in BLN to calculate an indicator}
  \item{variable}{Name of a variable for a specific indicator}
  \item{ess_theme}{Highest order grouping of indicators}
  \item{sub_group}{Second order grouping of indicators}
}
}
\keyword{datasets}
