% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_gw_pesticide.R
\name{bln_wat_pesticide}
\alias{bln_wat_pesticide}
\title{Function to calculate and evaluate the pesticide retention in soils in view
 of water purification for groundwater water quality}
\usage{
bln_wat_pesticide(
  ID,
  B_LU_BRP,
  B_SOILTYPE_AGR,
  A_CLAY_MI,
  A_SAND_MI,
  A_SILT_MI,
  A_SOM_LOI,
  M_GREEN,
  M_MECHWEEDS,
  M_PESTICIDES_DST
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional, option: yes or no)}

\item{M_MECHWEEDS}{(boolean) measure. Use of mechanical weed protection (option: yes or no)}

\item{M_PESTICIDES_DST}{(boolean) measure. Use of DST for pesticides (option: yes or no)}
}
\description{
Function to calculate and evaluate the pesticide retention in soils in view
 of water purification for groundwater water quality
}
