% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_helpers.R
\name{blnAssertLower}
\alias{blnAssertLower}
\title{Aid to assert lower values}
\usage{
blnAssertLower(x)
}
\arguments{
\item{x}{a value obtained by `bln_parms[code == 'variable_name', value_min]`}
}
\value{
x when it is numeric and not NA, else returns -Inf
}
\description{
Checks whether x is numeric and not NA. If so, returns x, else returns -Inf
}
\examples{
\dontrun{
BLN:::blnAssertLower(2)
BLN:::blnAssertLower('')
BLN:::blnAssertLower(NA_real_)

# example in context
A_SOM_LOI <- 5
arg.length = length(A_SOM_LOI)
checkmate::assert_numeric(A_SOM_LOI,
any.missing = FALSE, len = arg.length,
lower = BLN:::blnAssertLower(BLN::bln_parms[code == 'A_SOM_LOI', value_min])
)
}
}
\keyword{internal}
