% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_helpers.R
\name{checkvar}
\alias{checkvar}
\title{Select the correct boundary values (min and max) for a given parameter}
\usage{
checkvar(parm)
}
\arguments{
\item{parm}{(numeric) The bln parameter}
}
\value{
The min and max for the selected numeric parameter and options for categorial parameter
}
\description{
This function selects the correct boundary values (min and max) or possible options for a given parameter
}
