% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_helpers.R
\name{funArgs}
\alias{funArgs}
\title{Get vector of function arguments}
\usage{
funArgs(functionName, whichArgs = "all")
}
\arguments{
\item{functionName}{Quoted name of a function}

\item{whichArgs}{Select whether you want to return all arguments ('all'),
arguments without defaults ('required') or arguments with defaults ('optional'). Default is 'all'.}
}
\value{
a character vector of function argument names
}
\description{
Get vector of function arguments
}
\examples{
funArgs('sd')
funArgs('sd', whichArgs = 'required')
funArgs('sd', whichArgs = 'optional')

}
