% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_helpers.R
\name{funArgsV}
\alias{funArgsV}
\title{Get a vector of function arguments for multiple functions}
\usage{
funArgsV(functionNameVector, whichArgs = "all")
}
\arguments{
\item{functionNameVector}{A character vector of function names.}

\item{whichArgs}{Select whether you want to return all arguments ('all'),
arguments without defaults ('required') or arguments with defaults ('optional').
Default is 'all'.}
}
\value{
A character vector of unique function argument names.
}
\description{
This function is a wrapper around funArgs. It takes a vector of
function names and returns a single character vector with unique argument names.
}
\examples{
funArgsV(c('sd', 'mean'))
funArgsV(c('sd', 'mean'), whichArgs = 'required')

}
