% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_plots.R
\name{plot_bln_boxplot}
\alias{plot_bln_boxplot}
\title{Plot boxplots for BLN indicators and scores for a group of fields}
\usage{
plot_bln_boxplot(
  dt,
  pesd = "crop",
  ptitle = NULL,
  psubtitle = NULL,
  plegend = TRUE,
  pgroup = NULL,
  lpos = c(0.2, 0.8),
  ptype = "scores"
)
}
\arguments{
\item{dt}{(data.table) data.table with results from BLN calculations.}

\item{pesd}{(vector) a vector to select all indicators per ESD to be plotted. Options: crop, water, carbon, nutrient, or scores}

\item{ptitle}{(character) a title to be added to the figure. Default is NULL.}

\item{psubtitle}{(character) a subtitle to be added to the figure. Default is NULL.}

\item{plegend}{(boolean) a boolean argument to show legend. Default is TRUE}

\item{pgroup}{(character) a variable to facetwrap the boxplot}

\item{lpos}{(vector) a two dimensional vector to position the legend inside the plot. Default c(0.2,0.8).}

\item{ptype}{(character) a figure type. options: scores, knelpunten}
}
\description{
Plot boxplots for BLN indicators and scores for a group of fields
}
