% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_plots.R
\name{plot_bln_map}
\alias{plot_bln_map}
\title{Plot a BLN score or indicator on a map}
\usage{
plot_bln_map(
  dt.scores,
  dt.region = NULL,
  parm,
  pbreak = NULL,
  plabel = NULL,
  pregio_level = NULL,
  pregio_label = NULL,
  ptitle = NULL,
  psubtitle = NULL,
  rev = FALSE,
  lpos = c(0.2, 0.8),
  ptype = "map"
)
}
\arguments{
\item{dt.scores}{(data.table) data.table with results from BLN calculations. Includes a spatial geometry.}

\item{dt.region}{(data.table) data.table with a spatial geometry for a given region to be plot on top of the fields. Optional.}

\item{parm}{(character) one of the BLN indicator or scores to be plotted}

\item{pbreak}{(vector) a vector to categorize the input in classes. Default is NULL.}

\item{plabel}{(vector) a vector with label names used for the legend. Default is NULL.}

\item{pregio_level}{(vector) a vector with names used for the regio stats. Order defines plotting order. Default is NULL.}

\item{pregio_label}{(vector) a vector with label names used for the regio stats. Default is NULL.}

\item{ptitle}{(character) a title to be added to the figure. Default is NULL.}

\item{psubtitle}{(character) a subtitle to be added to the figure. Default is NULL.}

\item{rev}{(boolean) a boolean argument to reverse the color mapping. Default is FALSE.}

\item{lpos}{(vector) a two dimensional vector to position the legend inside the plot. Default c(0.2,0.8).}

\item{ptype}{(character) a figure type. options: map, mapplusstats}
}
\description{
Plot a BLN score or indicator on a map
}
