% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_plots.R
\name{plot_bln_rothc_ts}
\alias{plot_bln_rothc_ts}
\title{Function to plot simple RothC simulation time series for a single field}
\usage{
plot_bln_rothc_ts(dt, ptitle = NULL, psubtitle = NULL, lpos = c(0.2, 0.8))
}
\arguments{
\item{dt}{(data.table) data.table with results from BLN RothC simulation, calculated with bln_rothc_field}

\item{ptitle}{(character) a title to be added to the figure. Default is NULL.}

\item{psubtitle}{(character) a subtitle to be added to the figure. Default is NULL.}

\item{lpos}{(vector) a two dimensional vector to position the legend inside the plot. Default c(0.2,0.8).}
}
\description{
Function to plot simple RothC simulation time series for a single field
}
