% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_rothc.R
\name{rothc_scenario}
\alias{rothc_scenario}
\title{Function to prepare inputs for various scenarios with RothC}
\usage{
rothc_scenario(B_LU_BRP, scen)
}
\arguments{
\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{scen}{(character) possible scenarios. Include BAU, CLT, BAUIMPR and ALL}
}
\description{
Function to prepare inputs for various scenarios with RothC
}
