% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln.R
\name{Binomial-logit-normal}
\alias{Binomial-logit-normal}
\alias{bln}
\alias{dbln}
\alias{dblnx}
\alias{dblnpp}
\alias{dblnxpp}
\alias{pbln}
\alias{qbln}
\alias{rbln}
\title{The Binomial-Logit-Normal Distribution}
\usage{
dbln(x, size, mean = 0, sd = 1)

dblnx(x, size, mean = 0, sd = 1, drop = FALSE)

dblnpp(x, size, mean = 0, sd = 1)

dblnxpp(x, size, mean = 0, sd = 1)

pbln(q, size, mean = 0, sd = 1)

qbln(p, size, mean = 0, sd = 1)

rbln(n, size, mean = 0, sd = 1)
}
\arguments{
\item{x, q}{A vector of quantiles}

\item{size}{Number of trials}

\item{mean}{A vector of means}

\item{sd}{A vector of standard deviations}

\item{p}{A vector of probabilities}

\item{n}{Number of observations, if \code{length(x = n) > 1},
the length is taken to be the number required}
}
\value{
\code{dbln} gives the density

\code{dblnpp} gives the density using C++

\code{dblnxpp} gives the exact density using C++

\code{pbln} gives the distribution function

\code{qbln} gives the quantile function

\code{rbln} generates random deviates
}
\description{
Density, distribution function, quantile function, and random generation for
the binomial-logit-normal distribution with mean equal to \code{mean} and
standard deviation equal to \code{sd}, as well as distribution parameter \code{size}.
}
\references{
\code{dblnx} gives the exact density
}
