% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_phosphor.R
\name{bln_c_posphor}
\alias{bln_c_posphor}
\title{Calculate the phosphate availability index in the Netherlands}
\usage{
bln_c_posphor(B_LU, A_P_AL = NA_real_, A_P_CC = NA_real_, A_P_WA = NA_real_)
}
\arguments{
\item{B_LU}{(numeric) The crop code}

\item{A_P_AL}{(numeric) The P-content of the soil extracted with ammonium lactate}

\item{A_P_CC}{(numeric) The P-content of the soil extracted with CaCl2}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water}
}
\value{
The phosphate availability index in the Netherlands estimated from extractable soil P fractions. A numeric value.
}
\description{
This function calculates the phosphate availability.
}
\examples{
bln_c_posphor(B_LU = 265, A_P_AL = 45, A_P_CC = 2.5)
bln_c_posphor(B_LU = c(265,1019),A_P_AL = c(35,54),A_P_CC = c(2.5,4.5), A_P_WA = c(35,65))

}
