% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_csaturation.R
\name{bln_clim_csat}
\alias{bln_clim_csat}
\title{Function to evaluate the carbon saturation via max C saturation potential estimated via ML modelling}
\usage{
bln_clim_csat(B_LU_BRP, A_SOM_LOI, A_CLAY_MI, A_SOM_LOI_MLMAX = NA_real_)
}
\arguments{
\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SOM_LOI_MLMAX}{(numeric) maximum value for the soil organic matter content of the soil estimated with ML model of Ros et al. (2024)}
}
\description{
Function to evaluate the carbon saturation via max C saturation potential estimated via ML modelling
}
