% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_somers.R
\name{bln_clim_somers}
\alias{bln_clim_somers}
\title{Function to evaluate the carbon saturation via SOMERS simulation on peat soils}
\usage{
bln_clim_somers(
  ID,
  B_SOILTYPE_AGR,
  A_SOM_LOI,
  B_SOMERS_BC,
  B_DRAIN_SP,
  B_DRAIN_WP,
  B_DRAIN_SP_CHANGE = 0.2
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{B_SOMERS_BC}{(integer) The base combination of SOMERS peat soil classification (varies between 1 and 290)}

\item{B_DRAIN_SP}{(numeric) the drooglegging of a field in summer (difference field height and ditch level, in meters)}

\item{B_DRAIN_WP}{(numeric) the drooglegging of a field in winter (difference field height and ditch level, in meters)}

\item{B_DRAIN_SP_CHANGE}{(numeric) the decrease in drooglegging of a field in summer (in meters). Allowed decrease varies from 0 to 0.5m.}
}
\description{
Function to evaluate the carbon saturation via SOMERS simulation on peat soils
}
