% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_crops}
\alias{bln_crops}
\title{Linking table between crops and different functions in OBIC}
\format{
A data.frame with 465 rows and 3 columns:
\describe{
  \item{crop_code}{The BRP gewascode of the crop}
  \item{crop_name}{The name of the crop, in lower case}
  \item{crop_cat1}{Classification of crop per land use type (arable, maize, grass, nature)}
  \item{bln_country}{the country name where the crop codes are applicable}
}
}
\usage{
bln_crops
}
\description{
This table helps to link the different crops in the OBIC functions with the crops selected by the user
}
\keyword{datasets}
