% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_tables.R
\docType{data}
\name{bln_input_description}
\alias{bln_input_description}
\title{A dataset with all input variables required by the BLN package}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 5 columns.
}
\usage{
bln_input_description
}
\description{
This table contains all input parameters needed to run the BLN package
}
\details{
\describe{
  \item{code}{The input parameter's name}
  \item{parameter}{A description of the parameter}
  \item{data_type}{Type of data, either boolean, numeric, integer, or character}
  \item{value_min}{Minimum value for a numeric or integer parameter}
  \item{value_max}{Maximum value for a numeric or integer parameter}
}
}
\keyword{datasets}
