% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_rothc_events.R
\name{bln_rothc_event_crop}
\alias{bln_rothc_event_crop}
\title{Calculate the crop rotation related C inputs of a field on monthly basis for the Netherlands}
\usage{
bln_rothc_event_crop(crops, A_CLAY_MI)
}
\arguments{
\item{crops}{(data.table) Table with crop rotation, crop management measures, year and potential Carbon inputs.}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function determines how much Carbon enters the soil throughout the year given the crop rotation plan.
}
