% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_rothc.R
\name{bln_rothc_multicore}
\alias{bln_rothc_multicore}
\title{Function to evaluate the carbon saturation via RothC simulation on multiple fields using multicore processing}
\usage{
bln_rothc_multicore(
  ID,
  B_LU_BRP,
  B_GWL_GLG,
  A_SOM_LOI,
  A_CLAY_MI,
  scen,
  simyears = 50,
  quiet = TRUE
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{B_GWL_GLG}{(numeric) The lowest groundwater level averaged over the most dry periods in 8 years in cm below ground level}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{A_CLAY_MI}{(numeric) value for the clay content of the soil}

\item{scen}{(character) scenarios to be simulated. Options include BAU, BAUIMPR,CLT and ALL.}

\item{simyears}{(integer) value for the amount of years to simulate, default is 50 years}

\item{quiet}{(boolean) showing progress bar for calculation RothC C-saturation for each field}
}
\value{
A data.table with an ID column and one column for each scenario selected
with the scen argument which contain the predicted soil organic matter content
of that scenario at the end of the simulation.
}
\description{
Function to evaluate the carbon saturation via RothC simulation on multiple fields using multicore processing
}
\examples{
\donttest{
bln_rothc_multicore(ID = 1:10,
B_LU_BRP = rep(c(3732,265),5),
B_GWL_GLG = rep(95, 10),
A_SOM_LOI = seq(1,15,length.out = 10),
A_CLAY_MI = rep(4.5,10),
scen = c('BAU','ALL'),
quiet = FALSE)
}

}
