% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_helpers.R
\name{blnAssertLower}
\alias{blnAssertLower}
\title{Aid to assert lower values}
\usage{
blnAssertLower(x)
}
\arguments{
\item{x}{a value obtained by `bln_parms[code == 'variable_name', value_min]`}
}
\value{
x when it is numeric and not NA, else returns -Inf
}
\description{
Checks whether x is numeric and not NA. If so, returns x, else returns -Inf
}
\examples{
blnAssertLower(2)
blnAssertLower('')
blnAssertLower(NA_real_)

}
