% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_helpers.R
\name{blnAssertUpper}
\alias{blnAssertUpper}
\title{Aid to assert upper values}
\usage{
blnAssertUpper(x)
}
\arguments{
\item{x}{a value obtained by `bln_parms[code == 'variable_name', value_max]`}
}
\value{
x when it is numeric and not NA, else returns Inf
}
\description{
Checks whether x is numeric and not NA. If so, returns x, else returns Inf
}
\examples{
blnAssertUpper(2)
blnAssertUpper('')
blnAssertUpper(NA_real_)

# example in context
A_SOM_LOI <- 5
checkmate::assert_numeric(A_SOM_LOI,
any.missing = FALSE, len = arg.length,
lower = BLN::bln_parms[code == 'A_SOM_LOI', value_min],
upper = blnAssertUpper(BLN::bln_parms[code == 'A_SOM_LOI', value_max])
)

}
