% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bln_clim_rothc.R
\name{rothc_initialise}
\alias{rothc_initialise}
\title{Function for initializing the RothC model}
\usage{
rothc_initialise(B_LU_BRP, A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{B_LU_BRP}{(numeric) value of the BRP crop code}

\item{A_SOM_LOI}{(numeric) value for the soil organic matter content of the soil}

\item{A_CLAY_MI}{(numeric) value for the clay content of the soil}
}
\value{
A data table with the fractions IOM, DPM, RPM and BIO
}
\description{
Function for initializing the RothC model
}
\examples{
rothc_initialise(B_LU_BRP = 265,
                 A_SOM_LOI = 15,
                 A_CLAY_MI = 10)

}
